+++ source/extensions/omni.isaac.benchmarks/omni/isaac/benchmarks/tests/__init__.py
@@ -17,6 +17,7 @@ from .test_benchmark_real_time_factor import *
 from .test_benchmark_robots_carter import *
 from .test_benchmark_robots_o3dyn import *
 from .test_benchmark_rtx_lidar import *
+from .test_benchmark_rtx_radar import *
 from .test_benchmark_scene_generation import *
 from .test_benchmark_sdg_generation import *